pyproject.toml
[tool.poetry]
name = "finellm"
version = "0.1.0"
description = "POC Agent IA de Triage Medical - Centre Hospitalier Saint-Aurelien"
authors = ["marintosti12 <marintosti12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

# Core ML
torch = "^2.5.0"
transformers = "^4.47.0"
accelerate = "^1.2.0"
peft = "^0.14.0"
trl = "^0.13.0"
bitsandbytes = "^0.45.0"

# Datasets
datasets = "^3.2.0"

# Evaluation
evaluate = "^0.4.0"
scikit-learn = "^1.6.0"
rouge-score = "^0.1.2"

# Deployment
vllm = "^0.6.0"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"

# Utils
wandb = "^0.19.0"
pyyaml = "^6.0"
pandas = "^2.2.0"
tqdm = "^4.67.0"
presidio-analyzer = "^2.2"
presidio-anonymizer = "^2.2"
spacy = "^3.7"
matplotlib = "^3.10.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.8.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
